Revert the sillyness from 128f655f which made battery get return a int16_t. it returns an int8_t, thank you very much. (should probably be some fixed/floating point representation).

+++ hax.h
@@ -115,8 +115,8 @@ uint16_t analog_adc_get(index_t apin_index);
 bool digital_oi_get(index_t oi_index);
 int8_t analog_oi_get(index_t oi_index);
 
-/* TODO: Document this method. */
-int16_t battery_get(void);
+/* returns something resembling a battery voltage */
+uint8_t battery_get(void);
 
 /*
  * MOTOR AND SERVO OUTPUTS
